+++ tests/test_extensions/test_blocks/test_general_blocks.py
@@ -384,6 +384,78 @@ class TestAttributes(util.MdCase):
         )
 
 
+class TestNestedBlocksAndLists(util.MdCase):
+    """Test Nested blocks and lists."""
+
+    extension = ['pymdownx.blocks.tab', 'pymdownx.blocks.html']
+    extension_configs = {
+        'pymdownx.blocks.tab': {'alternate_style': True}
+    }
+
+    def test_nested_blocks_in_lists(self):
+        """Test a nested blocks case with lists."""
+
+        self.check_markdown(
+            R"""
+            //// html | div.my-div
+
+            - List
+
+                - List
+
+                    /// tab | TEST1
+
+                    Content
+
+                    ///
+
+                    /// tab | TEST2
+
+                    - A list
+
+                        Paragraph
+
+                            Code
+
+                    ///
+
+            ////
+            """,
+            """
+            <div class="my-div">
+            <ul>
+            <li>
+            <p>List</p>
+            <ul>
+            <li>
+            <p>List</p>
+            <div class="tabbed-set tabbed-alternate" data-tabs="1:2"><input checked="checked" id="__tabbed_1_1" name="__tabbed_1" type="radio" /><input id="__tabbed_1_2" name="__tabbed_1" type="radio" /><div class="tabbed-labels"><label for="__tabbed_1_1">TEST1</label><label for="__tabbed_1_2">TEST2</label></div>
+            <div class="tabbed-content">
+            <div class="tabbed-block">
+            <p>Content</p>
+            </div>
+            <div class="tabbed-block">
+            <ul>
+            <li>
+            <p>A list</p>
+            <p>Paragraph</p>
+            <pre><code>Code
+            </code></pre>
+            </li>
+            </ul>
+            </div>
+            </div>
+            </div>
+            </li>
+            </ul>
+            </li>
+            </ul>
+            </div>
+            """,  # noqa: E501
+            True
+        )
+
+
 class TestBlocksMdInHTML(util.MdCase):
     """Test blocks with `md_in_html`."""
 
@@ -391,7 +463,6 @@ class TestBlocksMdInHTML(util.MdCase):
     extension_configs = {
         'pymdownx.blocks.tab': {'alternate_style': True}
     }
-
 
     def test_md_in_html_inserted_correctly(self):
         """Test that `md_in_html` inserts under the correct target."""